Generalize properties framework
Generalize the properties framework so that more information can be passed to the UI beyond progress, error, and complete. Add VortexProperty class that includes four properties - status, progress, error, and complete. Status updates will update the label in the importer UI.

Set progress based on the number of ImportableUnits that have been processed. This includes a refactor of the SerialBatchImporter which previously used a very memory-intensive approach and did not use ImportableUnits.

Add a Stopwatch utility class for code re-use.