Update handling and tests for GPM IMERG data
A notable issue was found in NetcdfDataReader::getFloatArray where the previous approach for retrieving a float array resulting in transposed values for the GPM dataset. Updated the method to use the Array::get1DJavaArray method which returns data in the expected order starting from the top left corner of the grid and going row by row. Added unit tests to capture regression for both hourly and daily GPM IMERG data.