Implement TemporalDataReader
Add TemporalDataReader class to read from the specified data source or DataReader, and return a VortexGrid with the data for the specified time range. The calculations for the resulting data based on the time range is done via GridDataProcessor. Different calculations are used depending on the data type (instantaneous, average, or accumulation). BufferedDataReader is used to lazily read in data from the data source without causing out of memory error.