Represent cell size units as a Quantity
This fixes a bug where the UI indicated that the cell size units were always in meters but in actuality the cell size units would correspond to the target CRS specified in the resampling parameters/GDAL warp.

Add a combo box selection for cell size units in the importer UI to reduce confusion about which cell size units are being applied.