Fix TemporalDataReader::getGridsForPeriod
The previous implementation of the getGridsForPeriod's filtering logic
had a flaw where it would stop early if the available grids were
disjoint (not contiguous). This caused it to miss relevant grids and
return an incomplete data set.

The new implementation addresses this issue by considering all the
available grids, even if they are not contiguous. It will now correctly
identify and include all the relevant grids in the final result,
providing a more accurate and complete data set.

An example scenario:
- Query for period 1 - 5
- Available grids: [1-2], [3-4], [4-5]
- Previous implementation would only choose [1-2]
- New implementation chooses all grids: [1-2], [3-4], [4-5]

Resolves: HMS-4275