Remove special type designation for preciprate
Remove the special file type designation for preciprate datasets. There is no way to distinguish the 2-minute data from the 5-minute data by filename alone. Without the special type designation, the workflow is to import the dataset (the start/end times will be the same), then time-shift the start or end times.