Add message panel to wizards
In all the Vortex wizards, add a message panel to communicate status in addition to progress. This allows note/error/warning messages to be passed during processing.

The implementation for most wizards is similar. The Normalizer wizard is different in that it uses the java logging framework to pass messages as opposed to PropertyChangeSupport. To be determined if one approach is preferable of the other.